// src/app/core/modules/openapi/model/user-body.ts
/**
 * Adres API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Role } from './role';


export interface UserBody { 
    username: string;
    /**
     * base64 encode password
     */
    password?: string;
    email: string;
    phone?: string;
    accountNonExpired?: boolean;
    accountNonLocked?: boolean;
    credentialsNonExpired?: boolean;
    enabled?: boolean;
    roles?: Array<Role>;
}